% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawler.R
\name{getAirlineCode}
\alias{getAirlineCode}
\title{getAirlineCode}
\usage{
getAirlineCode(origin = "TPE", destination = "SEA")
}
\arguments{
\item{origin}{the code of your original airport}

\item{destination}{the code of your destination airport}
}
\description{
Get Airline Codes
}
\examples{
getAirlineCode("TPE","SFO")
}

