% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawler.R
\name{getFlightsData}
\alias{getFlightsData}
\title{getFlightsData}
\usage{
getFlightsData(fareType = "C", startDate = "2016/09/03",
  endDate = "2016/09/03", journeyType = "來回", originCode = "TN",
  destinationCode = "SEA", bankName = "全部")
}
\description{
Get Flights Data
}
\examples{
getFlightsData(fareType="Y") \%>\% View
}

